+++ test/conversion_test.exs
@@ -11,10 +11,10 @@ defmodule Cldr.Unit.Conversion.Test do
     end
   end
 
-  # Test #187 is reduced in rounding by one digit
-  # in order for the match to work.
+  # Test 185 is a Beaufort conversion. That needs special handling we
+  # don't have yet. And the test case data is unexpected.
 
-  @just_outside_tolerance []
+  @just_outside_tolerance [185]
 
   for t <- ConversionData.conversions(), t.line not in @just_outside_tolerance do
     test "##{t.line} [Float] that #{t.from} converted to #{t.to} is #{inspect(t.result)}" do
@@ -35,7 +35,7 @@ defmodule Cldr.Unit.Conversion.Test do
     end
   end
 
-  @just_outside_tolerance_decimal []
+  @just_outside_tolerance_decimal [185]
 
   @one_thousand Decimal.new(1000)
   for t <- ConversionData.conversions(), t.line not in @just_outside_tolerance_decimal do
@@ -78,5 +78,14 @@ defmodule Cldr.Unit.Conversion.Test do
 
   test "conversion where base units don't match but unit categories do" do
     assert {:ok, _} = Cldr.Unit.convert(Cldr.Unit.new!(:joule, 1), "kilowatt_hour")
+  end
+
+  test "converting from/to beaufort" do
+    beaufort = Cldr.Unit.new!(:beaufort, 17)
+    meter_per_second = Cldr.Unit.convert!(beaufort, :meter_per_second) |> Cldr.Unit.Math.round(1)
+    beaufort_2 = Cldr.Unit.convert!(meter_per_second, :beaufort) |> Cldr.Unit.Math.round(0)
+
+    assert meter_per_second == Cldr.Unit.new!(:meter_per_second, "58.6")
+    assert beaufort_2 == Cldr.Unit.new!(:beaufort, 17)
   end
 end