+++ test/unit_math_test.exs
@@ -25,4 +25,11 @@ defmodule Cldr.Unit.Math.Test do
     float_unit = Cldr.Unit.new!(:meter, "12.34") |> Cldr.Unit.to_float_unit() |> Cldr.Unit.round(1)
     assert float_unit.value == 12.3
   end
+
+  test "Mult/2 and div/2 with a scalar" do
+    assert  Cldr.Unit.Math.div(Cldr.Unit.new!(:meter, 10), 2) == Cldr.Unit.new!(:meter, 5)
+    assert  Cldr.Unit.Math.mult(Cldr.Unit.new!(:meter, 10), 2) == Cldr.Unit.new!(:meter, 20)
+  end
 end
+
+